+++ app/Http/Controllers/RoleUsersController.php
@@ -27,29 +27,34 @@ class RoleUsersController extends Controller
      */
     public function index(Request $request)
     {
-        if ($request->searchField) {
-            $roleUsers = DB::table('role_user')
-                            ->select('role_user.*', 'roles.display_name', 'users.name')
-                            ->join('users', 'users.id', 'role_user.user_id')
-                            ->join('roles', 'roles.id', 'role_user.role_id')
-                            ->where('roles.display_name', 'like', '%'.$request->searchField.'%')
-                            ->orWhere('users.name', 'like', '%'.$request->searchField.'%')
-                            ->orWhere('users.email', 'like', '%'.$request->searchField.'%')
-                            ->orderBy('role_user.role_id', 'asc')
-                            ->paginate();
-        } else {
-            $roleUsers = DB::table('role_user')
-                            ->select('role_user.*', 'roles.display_name', 'users.name')
-                            ->join('users', 'users.id', 'role_user.user_id')
-                            ->join('roles', 'roles.id', 'role_user.role_id')
-                            ->orderBy('role_user.role_id', 'asc')
-                            ->paginate();
-        }
+        if (Auth::user()->canListarRoleUser()) {
+            if ($request->searchField) {
+                $roleUsers = DB::table('role_user')
+                                ->select('role_user.*', 'roles.display_name', 'users.name')
+                                ->join('users', 'users.id', 'role_user.user_id')
+                                ->join('roles', 'roles.id', 'role_user.role_id')
+                                ->where('roles.display_name', 'like', '%'.$request->searchField.'%')
+                                ->orWhere('users.name', 'like', '%'.$request->searchField.'%')
+                                ->orWhere('users.email', 'like', '%'.$request->searchField.'%')
+                                ->orderBy('role_user.role_id', 'asc')
+                                ->paginate();
+            } else {
+                $roleUsers = DB::table('role_user')
+                                ->select('role_user.*', 'roles.display_name', 'users.name')
+                                ->join('users', 'users.id', 'role_user.user_id')
+                                ->join('roles', 'roles.id', 'role_user.role_id')
+                                ->orderBy('role_user.role_id', 'asc')
+                                ->paginate();
+            }
 
-        return View('role_user.index', [
-                'roleUsers' => $roleUsers,
-                'fields' => $this->fields
-        ]);
+            return View('role_user.index', [
+                    'roleUsers' => $roleUsers,
+                    'fields' => $this->fields
+            ]);
+        } else {
+            Session::flash('error', __('messages.access_denied'));
+            return redirect()->back();   
+        } 
     }
 
     /**
@@ -60,17 +65,14 @@ class RoleUsersController extends Controller
     public function create()
     {
         if (Auth::user()->canCadastrarRoleUser()) {
-            $roles = Role::all();
-            $users = User::all();
-
             return View('role_user.create', [
-                'roles' => $roles,
-                'users' => $users
+                'roles' => Role::all(),
+                'users' => User::all()
             ]);
         } else {
-            Session::flash('error', env('ACCESS_DENIED_MSG'));
-            return redirect()->back();
-        }
+            Session::flash('error', __('messages.access_denied'));
+            return redirect()->back();   
+        } 
     }
 
     /**
@@ -81,37 +83,36 @@ class RoleUsersController extends Controller
      */
     public function store(Request $request)
     {
-        $role = Role::find($request->role_id);
-        $user = User::find($request->user_id);
+        if (Auth::user()->canCadastrarRoleUser()) {
+            $role = Role::find($request->role_id);
+            $user = User::find($request->user_id);
 
-        $this->validate($request, [
-            'role_id' => 'required|numeric',
-            'user_id' => ['required', 'numeric', new ValidRoleUser($role, $user)],
-        ]);
+            $this->validate($request, [
+                'role_id' => 'required|numeric',
+                'user_id' => ['required', 'numeric', new ValidRoleUser($role, $user)],
+            ]);
 
-        try {
-            $roleUser = new RoleUser($request->all());
-            $roleUser->user_type = 'App\User';
-            
-            if ($roleUser->save()) {
-                Session::flash('success', 'Role vs User association successfull created');
-                return redirect()->action('RoleUsersController@index');
+            try {
+                $roleUser = new RoleUser($request->all());
+                $roleUser->user_type = 'App\User';
+                
+                if ($roleUser->save()) {
+                    Session::flash('success', __('messages.create_success', [
+                        'model' => __('role_user'),
+                        'name' => $roleUser->user->name.' - '.$roleUser->role->display_name
+                    ]));
+                    return redirect()->action('RoleUsersController@index');
+                }
+            } catch (\Exception $e) {
+                Session::flash('error', __('messages.exception', [
+                    'exception' => $e->getMessage()
+                ]));
+                return redirect()->back->withInput();
             }
-        } catch (\Exception $e) {
-            Session::flash('error', 'Oops, have one error...'.$e->getMessage());
-            return redirect()->back()->withInput();
+        } else {
+            Session::flash('error', __('messages.access_denied'));
+            return redirect()->back();   
         }
-    }
-
-    /**
-     * Display the specified resource.
-     *
-     * @param  \App\RoleUser  $roleUser
-     * @return \Illuminate\Http\Response
-     */
-    public function show(RoleUser $roleUser)
-    {
-        //
     }
 
     /**
@@ -122,17 +123,16 @@ class RoleUsersController extends Controller
      */
      public function edit(RoleUser $roleUser)
     {        
-        //dd($roleUser);
-        $users = User::all();
-        $roles = Role::all();
-        
-        $roleUser = RoleUser::find($roleUser->id);
-
-        return View('role_user.edit', [
-            'users' => $users,
-            'roles' => $roles,
-            'roleUser' => $roleUser
-        ]);
+        if (Auth::user()->canAlterarRoleUser()) {
+            return View('role_user.edit', [
+                'users' => User::all(),
+                'roles' => Role::all(),
+                'roleUser' => $roleUser
+            ]);
+        } else {
+            Session::flash('error', __('messages.access_denied'));
+            return redirect()->back();   
+        }
     }
 
     /**
@@ -144,28 +144,36 @@ class RoleUsersController extends Controller
      */
     public function update(Request $request, RoleUser $roleUser)
     {
-        $role = Role::find($request->role_id);
-        $user = User::find($request->user_id);
+        if (Auth::user()->canAlterarRoleUser()) {
+            $role = Role::find($request->role_id);
+            $user = User::find($request->user_id);
 
-        $this->validate($request, [
-            'role_id' => 'numeric|required',
-            'user_id' => ['required', 'numeric', new ValidRoleUser($role, $user)],
-        ]);
+            $this->validate($request, [
+                'role_id' => 'numeric|required',
+                'user_id' => ['required', 'numeric', new ValidRoleUser($role, $user)],
+            ]);
 
-        try {
-            $roleUser = RoleUser::find($roleUser->id);
-            $roleUser->user_id = $request->user_id;
-            $roleUser->role_id = $request->role_id;
+            try {
+                $roleUser = RoleUser::find($roleUser->id);
+                $roleUser->user_id = $request->user_id;
+                $roleUser->role_id = $request->role_id;
 
-            if ($roleUser->save()) {
-                Session::flash('success', 'Role vs User association successfull updated');
-                return redirect()->action('RoleUsersController@index');
-            } else {
-                throw new \Exception('Role User cannot be saved!');
+                if ($roleUser->save()) {
+                    Session::flash('success', __('messages.update_success', [
+                        'model' => __('role_user'),
+                        'name' => $roleUser->user->name.' - '.$roleUser->role->display_name
+                    ]));
+                    return redirect()->action('RoleUsersController@index');
+                } 
+            } catch (\Exception $e) {
+                Session::flash('error', __('messages.exception', [
+                    'exception' => $e->getMessage()
+                ]));
+                return redirect()->back->withInput();
             }
-        } catch (\Exception $e) {
-            Session::flash('error', 'Oops, have one error...'.$e->getMessage());
-            return redirect()->back()->withInput();
+        } else {
+            Session::flash('error', __('messages.access_denied'));
+            return redirect()->back();   
         }
     }
 
@@ -177,16 +185,33 @@ class RoleUsersController extends Controller
      */
     public function destroy(RoleUser $roleUser)
     {
-        try {
-            $roleUser = RoleUser::find($roleUser->id);
-            if ($roleUser->delete()) {
-                Session::flash('success', 'Role vs User '.$roleUser->user->name.' - '.$roleUser->role->display_name.' successfull removed.');
-                
+        if (Auth::user()->canAlterarRoleUser()) {
+            try {
+                $roleUser = RoleUser::find($roleUser->id);
+                if ($roleUser->delete()) {
+                    Session::flash('success', __('messages.delete_success', [
+                        'model' => __('role_user'),
+                        'name' => $roleUser->user->name.' - '.$roleUser->role->display_name
+                    ]));
+                    
+                    return redirect()->action('RoleUsersController@index');
+                }
+            } catch (\Exception $e) {
+                switch ($e->getCode()) {
+                    case 23000:
+                        Session::flash('error', __('messages.fk_exception'));
+                        break;
+                    default:
+                        Session::flash('error', __('messages.exception', [
+                            'exception' => $e->getMessage()
+                        ]));
+                        break;
+                }
                 return redirect()->action('RoleUsersController@index');
             }
-        } catch (\Exception $e) {
-            Session::flash('error', 'Oops, have one error...'.$e->getMessage());
-            return redirect()->action('RoleUsersController@index');
+        } else {
+            Session::flash('error', __('messages.access_denied'));
+            return redirect()->back();   
         }
     }
 }
